+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -1,7 +1,10 @@
 package tech.skot
 
+import kotlin.Int
+import kotlin.String
+
 public object Versions {
-  public const val skot: String = "1.2.42-ua"
+  public const val skot: String = "1.2.43-ua"
 
   public const val group: String = "com.github.useradgents.skot"