+++ viewcontract/src/commonMain/kotlin/tech/skot/core/di/CoreViewInjector.kt
@@ -108,6 +108,7 @@ interface CoreViewInjector {
         hiddenInitial: Boolean?,
         dropDownDisplayedInitial: Boolean,
         onInputText: (input: String?) -> Unit,
+        onFocusChange: ((hasFocus: Boolean) -> Unit)?,
         oldSchoolModeHint: Boolean,
     ): SKInputWithSuggestionsVC