+++ viewlegacy/src/androidMain/kotlin/tech/skot/core/di/CoreViewInjectorImpl.kt
@@ -184,6 +184,7 @@ class CoreViewInjectorImpl : CoreViewInjector {
         hiddenInitial: Boolean?,
         dropDownDisplayedInitial: Boolean,
         onInputText: (input: String?) -> Unit,
+        onFocusChange: ((hasFocus:Boolean) -> Unit)?,
         oldSchoolModeHint: Boolean
     ) = SKInputWithSuggestionsViewProxy(
         hint = hint,
@@ -195,6 +196,7 @@ class CoreViewInjectorImpl : CoreViewInjector {
         hiddenInitial = hiddenInitial,
         dropDownDisplayedInitial = dropDownDisplayedInitial,
         onInputText = onInputText,
+        onFocusChange = onFocusChange,
         oldSchoolModeHint = oldSchoolModeHint
     )